Downloading refactor

+++ borrentlib/chunk.go
@@ -6,21 +6,21 @@ import (
 )
 
 type ChunkPersister interface {
-	PersistChunk(idx int64, offset int64, data []byte) error
-	ReadChunkHash(idx uint64) ([]byte, error)
+	PersistChunk(idx uint32, offset uint32, data []byte) error
+	ReadChunkHash(idx uint32) ([]byte, error)
 }
 
 type SparseFileDiskChunkPersister struct {
 	fileName  string
 	file      *os.File
-	chunkSize int64
+	chunkSize uint32
 }
 
-func InitSparseFileDiskChunkPersister(fileName string, size uint64, chunkSize uint64) (p *SparseFileDiskChunkPersister, err error) {
+func InitSparseFileDiskChunkPersister(fileName string, size uint32, chunkSize uint32) (p *SparseFileDiskChunkPersister, err error) {
 	p = &SparseFileDiskChunkPersister{
 		fileName:  fileName,
 		file:      nil,
-		chunkSize: int64(chunkSize),
+		chunkSize: chunkSize,
 	}
 	f, err := os.Create(p.fileName)
 	if err != nil {
@@ -31,14 +31,14 @@ func InitSparseFileDiskChunkPersister(fileName string, size uint64, chunkSize ui
 	return
 }
 
-func (p *SparseFileDiskChunkPersister) PersistChunk(idx int64, offset int64, data []byte) error {
-	_, err := p.file.WriteAt(data, idx*p.chunkSize+offset)
+func (p *SparseFileDiskChunkPersister) PersistChunk(idx uint32, offset uint32, data []byte) error {
+	_, err := p.file.WriteAt(data, int64(idx*p.chunkSize+offset))
 	return err
 }
 
-func (p *SparseFileDiskChunkPersister) ReadChunkHash(idx uint64) ([]byte, error) {
+func (p *SparseFileDiskChunkPersister) ReadChunkHash(idx uint32) ([]byte, error) {
 	var buf = make([]byte, p.chunkSize)
-	_, err := p.file.ReadAt(buf, int64(idx)*p.chunkSize)
+	_, err := p.file.ReadAt(buf, int64(idx*p.chunkSize))
 
 	hash := sha1.Sum(buf)
 	return hash[:], err
